[UI] - made is so grading ui would display corrct and inccorrct answer properply  - made is so notfication of when the exit is avible would pop up  -staterted to work on anme for leader board

+++ ProjectMischief/Assets/Scripts/Gameplay/ArtManager.cs
@@ -81,7 +81,7 @@ public class ArtManager : MonoBehaviour
         correctChoices = correctChanges;
 
         float percentage = 1.0f;
-        percentage = (grade - gradePenalty) / gradeMax;
+        percentage = (float)(grade - gradePenalty) / (float)gradeMax;
         percentage *= 100.0f;
 
         if      (percentage >= minPercentageForS) return 'S';
